Move common code for AWS batch modules in module_utils/aws/batch
